experiments continued, creation of config-node working on import, not with existing flows, wip
... and the 'enablePolling' issue isn't sorted either.

However, there is a chance that this is on the route to a solution: Looking up
an existing config node (via RED.nodes.eachConfig()) does not seem to work
during registerNode() in victron-nodes.html: If the node does not exist yet (in
cast of pristine node-red state), we instantiate exactly one, that's working.
... but we would need a way to check if the config node exists otherwise for
this approach to work.